Create scrape_data_intosqlite3.py
This GitHub project is designed to automate the process of updating a SQLite3 database. The client has provided a database file that contains a table with two columns: 'Website' and 'html'. Initially, the 'Website' column is populated with URLs, while the 'html' column is empty.

The purpose of this project is to create a bot that can retrieve the homepage text of each URL and update the corresponding 'html' column in the database. By utilizing this automation process, the database will be enriched with the HTML content of the websites.

This project streamlines the task of populating the 'html' column by eliminating the need for manual retrieval and update of the webpage content. With the bot handling the process, the database can be kept up-to-date effortlessly.